+++ tests/test_utils.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import numpy as np
 import pytest
-from tomlkit import load
+import tomllib
 
 from hydromt_wflow import WflowModel, WflowSedimentModel
 from hydromt_wflow.utils import get_grid_from_config
@@ -156,11 +156,11 @@ def test_config_toml_grouping(tmpdir):
 
     dummy_model.write()
 
-    with open(tmpdir / "wflow_sbm.toml", "r") as file:
-        written_config = load(file)
+    with open(tmpdir / "wflow_sbm.toml", "rb") as file:
+        written_config = tomllib.load(file)
 
-    with open(TESTDATADIR / "grouped_model_config.toml") as file:
-        expected_config = load(file)
+    with open(TESTDATADIR / "grouped_model_config.toml", "rb") as file:
+        expected_config = tomllib.load(file)
 
     assert written_config == expected_config